drivers: thermal: Add support for THERMAL_MAX_LIMIT cdev
Thermal cooling device can define THERMAL_NO_LIMIT as a state in
devicetree if the default max or min state has to be used. If the user
has to apply a specific mitigation state from the max and min the state
has to be mentioned. If the device has multiple speed bins and if the
number of state changes, the max mitigation state will vary.

In order to apply max mitigation irrespective of the number of states,
add support for the THERMAL_MAX_LIMIT macro. Users can define any state
as a an offset from the THERMAL_MAX_LIMIT macro.

Change-Id: I1808f0c6caa2918ca978842a6dc348956d69590c
Signed-off-by: Ram Chandrasekar <rkumbako@codeaurora.org>