Revert "extcon: Modify extcon device to be created after driver data is set"
In case a driver simply allocates and registers an extcon device, we
will get operation on devres_head from the extcon device which whose
list hasn't been initialized yet.

  chg->extcon = devm_extcon_dev_allocate(&pdev->dev, smblib_extcon_cable);
  rc = devm_extcon_dev_register(&pdev->dev, chg->extcon);

We get a crash with the following message:

  [   32.649118] list_add corruption. prev is NULL.

It is unknown in which kernel version behavior was changed so the
original change works correctly.

This reverts commit 6e721f3ad0535b24f19a62420f4da95212cf069c.

Change-Id: Ia9f6280475a2313b9760161b483ec447eacd92eb